add access checks to crowdfundings user extensions and dedup testimonial
